refactor pagerduty pkg
