+++ cmd/pagerdutyoncallcollector/BUILD.bazel
@@ -7,8 +7,8 @@ go_library(
     importpath = "github.com/stebennett/squad-dashboard/cmd/pagerdutyoncallcollector",
     visibility = ["//visibility:private"],
     deps = [
-        "//pkg/pagerdutyrepository",
-        "//pkg/pagerdutyservice",
+        "//pkg/pagerduty/repo",
+        "//pkg/pagerduty/service",
         "//vendor/github.com/Netflix/go-env",
     ],
 )
