+++ pkg/pagerduty/service/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "service",
+    srcs = ["service.go"],
+    importpath = "github.com/stebennett/squad-dashboard/pkg/pagerduty/service",
+    visibility = ["//visibility:public"],
+    deps = ["//pkg/pagerduty/models"],
+)
