+++ pkg/pagerduty/service/service.go
@@ -1,4 +1,4 @@
-package pagerdutyservice
+package service
 
 import (
 	"encoding/json"
@@ -9,7 +9,7 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/stebennett/squad-dashboard/pkg/pagerdutymodels"
+	"github.com/stebennett/squad-dashboard/pkg/pagerduty/models"
 )
 
 type PagerDutyParams struct {
@@ -56,16 +56,16 @@ func NewPagerDutyService(httpClient *http.Client, params PagerDutyParams) *Pager
 	}
 }
 
-func (pds *PagerDutyService) GetOnCalls(since time.Time, until time.Time) ([]pagerdutymodels.OnCall, error) {
-	return pds.getOnCalls(0, 25, since, until, []pagerdutymodels.OnCall{})
+func (pds *PagerDutyService) GetOnCalls(since time.Time, until time.Time) ([]models.OnCall, error) {
+	return pds.getOnCalls(0, 25, since, until, []models.OnCall{})
 }
 
-func (pds *PagerDutyService) getOnCalls(offset int, limit int, since time.Time, until time.Time, result []pagerdutymodels.OnCall) ([]pagerdutymodels.OnCall, error) {
+func (pds *PagerDutyService) getOnCalls(offset int, limit int, since time.Time, until time.Time, result []models.OnCall) ([]models.OnCall, error) {
 	log.Printf("Fetching PagerDuty Oncalls with offset %d", offset)
 
 	req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("https://%s/oncalls", pds.params.BaseUrl), nil)
 	if err != nil {
-		return []pagerdutymodels.OnCall{}, err
+		return []models.OnCall{}, err
 	}
 
 	req.Header.Set("Accept", "application/vnd.pagerduty+json;version=2")
@@ -83,31 +83,31 @@ func (pds *PagerDutyService) getOnCalls(offset int, limit int, since time.Time, 
 
 	resp, err := pds.httpClient.Do(req)
 	if err != nil {
-		return []pagerdutymodels.OnCall{}, err
+		return []models.OnCall{}, err
 	}
 
 	body, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
-		return []pagerdutymodels.OnCall{}, err
+		return []models.OnCall{}, err
 	}
 
 	var oncallPageResponse PagerDutyOncallResponse
 	err = json.Unmarshal(body, &oncallPageResponse)
 	if err != nil {
-		return []pagerdutymodels.OnCall{}, err
+		return []models.OnCall{}, err
 	}
 
 	for _, onCallResp := range oncallPageResponse.OnCalls {
-		onCall := pagerdutymodels.OnCall{
-			User: pagerdutymodels.PagerDutyEntitySummary{
+		onCall := models.OnCall{
+			User: models.PagerDutyEntitySummary{
 				Id:   onCallResp.User.Id,
 				Name: onCallResp.User.Summary,
 			},
-			Schedule: pagerdutymodels.PagerDutyEntitySummary{
+			Schedule: models.PagerDutyEntitySummary{
 				Id:   onCallResp.Schedlule.Id,
 				Name: onCallResp.Schedlule.Summary,
 			},
-			EscalationPolicy: pagerdutymodels.PagerDutyEntitySummary{
+			EscalationPolicy: models.PagerDutyEntitySummary{
 				Id:   onCallResp.EscalationPolicy.Id,
 				Name: onCallResp.EscalationPolicy.Summary,
 			},
