--- pkg/pagerdutymodels/BUILD.bazel
@@ -1,8 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "pagerdutymodels",
-    srcs = ["pagerdutymodels.go"],
-    importpath = "github.com/stebennett/squad-dashboard/pkg/pagerdutymodels",
-    visibility = ["//visibility:public"],
-)
