--- pkg/pagerdutyservice/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "pagerdutyservice",
-    srcs = ["pagerdutyservice.go"],
-    importpath = "github.com/stebennett/squad-dashboard/pkg/pagerdutyservice",
-    visibility = ["//visibility:public"],
-    deps = ["//pkg/pagerdutymodels"],
-)
